Build demo for input groups
