+++ lib/NodeGraphEditor.stories.tsx
@@ -188,178 +188,6 @@ export const InputFields: Story = {
   },
 }
 
-export const InputGroups: Story = {
-  parameters: {
-    layout: 'fullscreen',
-  },
-  args: {
-    nodes: [
-      {
-        id: '1',
-        type: 'bsdf',
-        position: { x: 350, y: 100 },
-        data: {
-          __inputGroupsExpanded: ['Specular'],
-        },
-      },
-      {
-        id: '2',
-        type: 'number',
-        position: { x: 100, y: 100 },
-        data: {},
-      },
-      {
-        id: '3',
-        type: 'number',
-        position: { x: 100, y: 200 },
-        data: {},
-      },
-    ],
-    edges: [
-      {
-        id: 'e1',
-        source: '3',
-        sourceHandle: 'value',
-        target: '1',
-        targetHandle: 'strength',
-      },
-      {
-        id: 'e2',
-        source: '2',
-        sourceHandle: 'value',
-        target: '1',
-        targetHandle: 'anisotropicRotation',
-      },
-      {
-        id: 'e3',
-        source: '3',
-        sourceHandle: 'value',
-        target: '1',
-        targetHandle: 'anisotropic',
-      },
-    ],
-    config: {
-      valueTypes: {
-        string: {
-          name: 'String',
-          color: '#a1a1a1',
-          inputType: 'value',
-          defaultValue: '',
-        },
-        number: {
-          name: 'Number',
-          color: '#a1a1a1',
-          inputType: 'value',
-          defaultValue: '0.000',
-        },
-        boolean: {
-          name: 'Boolean',
-          color: '#a1a1a1',
-          inputType: 'checkbox',
-          defaultValue: true,
-        },
-        specularDistribution: {
-          name: 'Specular Distribution',
-          color: '#06b6d4',
-          inputType: 'options',
-          options: [
-            { name: 'GGX', value: 'ggx' },
-            { name: 'Beckmann', value: 'beckmann' },
-            { name: 'Phong', value: 'phong' },
-          ],
-          defaultValue: 'GET',
-        },
-      },
-      nodeGroups: {
-        default: {
-          name: 'Default',
-          color: '#2A652B',
-        },
-      },
-      nodes: {
-        number: {
-          group: 'default',
-          name: 'Number',
-          inputs: [
-            {
-              name: 'Value',
-              id: 'value',
-              valueType: 'number',
-              isConstant: true,
-            },
-          ],
-          outputs: [
-            {
-              name: 'Value',
-              id: 'value',
-              valueType: 'number',
-            },
-          ],
-        },
-        bsdf: {
-          group: 'default',
-          name: 'Principled BSDF',
-          inputs: [
-            {
-              name: 'Metallic',
-              id: 'metallic',
-              valueType: 'number',
-            },
-            {
-              name: 'Roughness',
-              id: 'roughness',
-              valueType: 'number',
-              defaultValue: '0.550',
-            },
-            {
-              name: 'IOR',
-              id: 'ior',
-              valueType: 'number',
-              defaultValue: '1.450',
-            },
-            {
-              name: 'Alpha',
-              id: 'alpha',
-              valueType: 'number',
-              defaultValue: '1.000',
-            },
-            {
-              name: 'Distribution',
-              id: 'distribution',
-              group: 'Specular',
-              valueType: 'specularDistribution',
-            },
-            {
-              name: 'IOR Level',
-              id: 'iorLevel',
-              group: 'Specular',
-              valueType: 'number',
-            },
-            {
-              name: 'Anisotropic',
-              id: 'anisotropic',
-              group: 'Specular',
-              valueType: 'number',
-            },
-            {
-              name: 'Anisotropic Rotation',
-              id: 'anisotropicRotation',
-              group: 'Specular',
-              valueType: 'number',
-            },
-            {
-              name: 'Strength',
-              id: 'strength',
-              group: 'Emission',
-              valueType: 'number',
-            },
-          ],
-        },
-      },
-    },
-  },
-}
-
 export const SelectedEdgeHighlighting: Story = {
   parameters: {
     layout: 'fullscreen',
