+++ package.json
@@ -83,6 +83,7 @@
     "vite-plugin-lib-inject-css": "^1.3.0"
   },
   "dependencies": {
+    "@uiw/react-color": "^2.0.6",
     "dagre": "^0.8.5",
     "lodash-es": "^4.17.21",
     "nanoid": "^5.0.4",
